Create `DeclarativeTableViewController` and remove `DeclarativeSection` logic from `TableViewController`
